Add pingInterval to config
Before source code was uploaded, the config contained "pingInterval" which was thought to be tied to a feature that I had removed upon releasing. Removing this config option meant that the bot would send packets to discord api every millisecond. "pingInterval" has since returned to the config fixing this bug that would cause the bot to be put offline.